Add the concept of ignored error lines to the autoclassification backend.
When an error line is verified with a null classification, it indicates
that it's a line not associated with any specific error but the result
of some other problem in the run (e.g. a failed test). As a result it is
not to be used as a target for future autoclassifications.

When a classification is given the special bug number 0, it indicates that
it's a log line that can always be ignored, but is not one specific intermittent
that has to be handled (e.g. a non-fatal error from a test, or output from some
other tool that shouldn't cause a run to fail)